requirements.txt
fastapi==0.79.0
deta==1.1.0
requests-futures==1.0.0
python-dotenv==0.20.0
